Se refactoriza BigDecimal por double en modelo y entidad

+++ domain/model/src/main/java/co/com/crediya/model/user/User.java
@@ -24,7 +24,7 @@ public class User {
     String email;
     String identityDocument;
     String phoneNumber;
-    BigDecimal baseSalary;
+    Double baseSalary;
     Role role;
     String password;
